server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic>=2.6
python-dotenv>=1.0.1
openai>=1.40.0
